+++ src/character.h
@@ -1834,6 +1834,26 @@ class Character : public Creature, public visitable<Character>
         int  run_cost( int base_cost, bool diag = false ) const;
         const pathfinding_settings &get_pathfinding_settings() const override;
         std::set<tripoint> get_path_avoid() const override;
+        /**
+         * Get all hostile creatures currently visible to this player.
+         */
+        std::vector<Creature *> get_hostile_creatures( int range ) const;
+
+        /**
+         * Returns all creatures that this player can see and that are in the given
+         * range. This player object itself is never included.
+         * The player character (g->u) is checked and might be included (if applicable).
+         * @param range The maximal distance (@ref rl_dist), creatures at this distance or less
+         * are included.
+         */
+        std::vector<Creature *> get_visible_creatures( int range ) const;
+        /**
+         * As above, but includes all creatures the player can detect well enough to target
+         * with ranged weapons, e.g. with infrared vision.
+         */
+        std::vector<Creature *> get_targetable_creatures( int range ) const;
+        /** Returns an enumeration of visible mutations with colors */
+        std::string visible_mutations( int visibility_cap ) const;
         player_activity get_destination_activity() const;
         void set_destination_activity( const player_activity &new_destination_activity );
         void clear_destination_activity();
@@ -2154,6 +2174,13 @@ class Character : public Creature, public visitable<Character>
         void suffer_from_artifacts();
         void suffer_from_stimulants( int current_stim );
         void suffer_without_sleep( int sleep_deprivation );
+        /**
+         * Check whether the other creature is in range and can be seen by this creature.
+         * @param critter Creature to check for visibility
+         * @param range The maximal distance (@ref rl_dist), creatures at this distance or less
+         * are included.
+         */
+        bool is_visible_in_range( const Creature &critter, int range ) const;
 
         // a cache of all active enchantment values.
         // is recalculated every turn in Character::recalculate_enchantment_cache
