+++ src/player.cpp
@@ -4757,23 +4757,6 @@ float player::hearing_ability() const
     return volume_multiplier;
 }
 
-std::string player::visible_mutations( const int visibility_cap ) const
-{
-    const std::vector<trait_id> &my_muts = get_mutations();
-    const std::string trait_str = enumerate_as_string( my_muts.begin(), my_muts.end(),
-    [visibility_cap ]( const trait_id & pr ) -> std::string {
-        const auto &mut_branch = pr.obj();
-        // Finally some use for visibility trait of mutations
-        if( mut_branch.visibility > 0 && mut_branch.visibility >= visibility_cap )
-        {
-            return colorize( mut_branch.name(), mut_branch.get_display_color() );
-        }
-
-        return std::string();
-    } );
-    return trait_str;
-}
-
 std::vector<std::string> player::short_description_parts() const
 {
     std::vector<std::string> result;
@@ -4805,51 +4788,6 @@ int player::print_info( const catacurses::window &w, int vStart, int, int column
 {
     mvwprintw( w, point( column, vStart++ ), _( "You (%s)" ), name );
     return vStart;
-}
-
-bool player::is_visible_in_range( const Creature &critter, const int range ) const
-{
-    return sees( critter ) && rl_dist( pos(), critter.pos() ) <= range;
-}
-
-std::vector<Creature *> player::get_visible_creatures( const int range ) const
-{
-    return g->get_creatures_if( [this, range]( const Creature & critter ) -> bool {
-        return this != &critter && pos() != critter.pos() && // TODO: get rid of fake npcs (pos() check)
-        rl_dist( pos(), critter.pos() ) <= range && sees( critter );
-    } );
-}
-
-std::vector<Creature *> player::get_targetable_creatures( const int range ) const
-{
-    return g->get_creatures_if( [this, range]( const Creature & critter ) -> bool {
-        bool can_see = sees( critter ) || sees_with_infrared( critter );
-        if( can_see )   //handles the case where we can see something with glass in the way or a mutation lets us see through walls
-        {
-            std::vector<tripoint> path = g->m.find_clear_path( pos(), critter.pos() );
-            for( const tripoint &point : path ) {
-                if( g->m.impassable( point ) ) {
-                    can_see = false;
-                    break;
-                }
-            }
-        }
-        bool in_range = std::round( rl_dist_exact( pos(), critter.pos() ) ) <= range;
-        // TODO: get rid of fake npcs (pos() check)
-        bool valid_target = this != &critter && pos() != critter.pos() && attitude_to( critter ) != Creature::Attitude::A_FRIENDLY;
-        return valid_target && in_range && can_see;
-    } );
-}
-
-std::vector<Creature *> player::get_hostile_creatures( int range ) const
-{
-    return g->get_creatures_if( [this, range]( const Creature & critter ) -> bool {
-        // Fixes circular distance range for ranged attacks
-        float dist_to_creature = std::round( rl_dist_exact( pos(), critter.pos() ) );
-        return this != &critter && pos() != critter.pos() && // TODO: get rid of fake npcs (pos() check)
-        dist_to_creature <= range && critter.attitude_to( *this ) == A_HOSTILE
-        && sees( critter );
-    } );
 }
 
 bool player::query_yn( const std::string &mes ) const
